requirements.txt
streamlit==1.30.*
pygame_gui